requirements.txt
flask==3.1.0
transformers==4.47.0
torch==2.5.1
spacy==3.8.4
PyPDF2==3.0.1
python-docx==1.1.2
nltk==3.9.1
langdetect==1.0.9
